requirements-tests.txt
-r requirements.txt
coverage==4.5.1
coveralls==1.3.0
nose==1.3.7
pylint==1.8.3
pytest==5.3.3
pytest-django==3.8.0
tox==2.7.0